✨ TraqUserExtension class
