+++ src/KnoqReminder.App/KnoqReminder.App/Helpers/Traq/TraqLogHelper.cs
@@ -10,40 +10,7 @@ static class TraqLogHelper
         return factory.CreateLogger(typeof(TraqLogHelper));
     }
 
-    public static async ValueTask<global::Traq.Models.UserDetail?> GetWithLogOnFailureAsync(
-        this global::Traq.Users.Item.WithUserItemRequestBuilder builder,
-        ILoggerFactory loggerFactory,
-        Action<RequestConfiguration<DefaultQueryParameters>>? requestConfiguration = null,
-        CancellationToken cancellationToken = default
-        )
-    {
-        try
-        {
-            return await builder.GetAsync(requestConfiguration, cancellationToken);
-        }
-        catch (ApiException ex) when (ex.ResponseStatusCode == (int)HttpStatusCode.NotFound)
-        {
-            var logger = CreateLogger(loggerFactory);
-            if (logger.IsEnabled(LogLevel.Error))
-            {
-                var userId = builder.ToGetRequestInformation(requestConfiguration).PathParameters["userId"];
-                if (userId is Guid uid)
-                {
-                    logger.LogError_FailedToGetUser_UserNotFound(uid);
-                }
-                else
-                {
-                    logger.LogError_FailedToGetUser(ex);
-                }
-            }
-        }
-        catch (ApiException ex)
-        {
-            var logger = CreateLogger(loggerFactory);
-            logger.LogError_FailedToGetUser(ex);
-        }
-        return null;
-    }
+    
 
     public static async ValueTask<global::Traq.Models.Message?> PostWithLogOnFailureAsync(
         this global::Traq.Channels.Item.Messages.MessagesRequestBuilder builder,
@@ -83,18 +50,6 @@ static class TraqLogHelper
 
 static partial class MessageLogger
 {
-    /// <summary>
-    /// User not found: {<paramref name="userId"/>}
-    /// </summary>
-    [LoggerMessage(Level = LogLevel.Error, Message = "User not found: {userId}")]
-    public static partial void LogError_FailedToGetUser_UserNotFound(this ILogger logger, Guid userId);
-
-    /// <summary>
-    /// Failed to get a traQ user.
-    /// </summary>
-    [LoggerMessage(Level =LogLevel.Error,Message ="Failed to get a traQ user.")]
-    public static partial void LogError_FailedToGetUser(this ILogger logger, Exception exception);
-
     /// <summary>
     /// Failed to send a message to traQ: channel not found: {<paramref name="channelId"/>}
     /// </summary>
