+++ src/KnoqReminder.App/KnoqReminder.App/Helpers/Traq/TraqUserExtension.cs
@@ -1,0 +1,90 @@
+using System.Net;
+using KnoqReminder.Utilities;
+using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Kiota.Abstractions;
+
+namespace KnoqReminder.App.Helpers.Traq;
+
+static class TraqUserExtension
+{
+    public static readonly KeyedMemoryCacheEntryOptions UserCacheOptions = new()
+    {
+        Key = "Traq:User",
+        Options = new() { AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(3) }
+    };
+
+    static ILogger CreateLogger(ILoggerFactory factory)
+    {
+        return factory.CreateLogger(typeof(TraqUserExtension));
+    }
+
+    public static async ValueTask<global::Traq.Models.UserDetail?> TryGetAsync(
+        this global::Traq.Users.Item.WithUserItemRequestBuilder builder,
+        ILoggerFactory loggerFactory,
+        Action<RequestConfiguration<DefaultQueryParameters>>? requestConfiguration = null,
+        CancellationToken cancellationToken = default
+        )
+    {
+        try
+        {
+            return await builder.GetAsync(requestConfiguration, cancellationToken);
+        }
+        catch (ApiException ex) when (ex.ResponseStatusCode == (int)HttpStatusCode.NotFound)
+        {
+            var logger = CreateLogger(loggerFactory);
+            if (logger.IsEnabled(LogLevel.Error))
+            {
+                var userId = builder.ToGetRequestInformation(requestConfiguration).PathParameters["userId"];
+                if (userId is Guid uid)
+                {
+                    logger.LogError_FailedToGetUser_UserNotFound(uid);
+                }
+                else
+                {
+                    logger.LogError_FailedToGetUser(ex);
+                }
+            }
+        }
+        catch (ApiException ex)
+        {
+            var logger = CreateLogger(loggerFactory);
+            logger.LogError_FailedToGetUser(ex);
+        }
+        return null;
+    }
+
+    public static async ValueTask<global::Traq.Models.UserDetail?> TryGetCachedAsync(
+        this global::Traq.Users.Item.WithUserItemRequestBuilder builder,
+        IMemoryCache cache,
+        ILoggerFactory loggerFactory,
+        Action<RequestConfiguration<DefaultQueryParameters>>? requestConfiguration = null,
+        CancellationToken cancellationToken = default
+        )
+    {
+        var userId = builder.ToGetRequestInformation(requestConfiguration).PathParameters["userId"];
+        if (userId is not Guid uid)
+        {
+            return null;
+        }
+        return await cache.GetOrCreateAsync(UserCacheOptions.GetMemoryCacheKey(uid), async entry =>
+        {
+            entry.SetOptions(UserCacheOptions.Options);
+            return await builder.TryGetAsync(loggerFactory, requestConfiguration, cancellationToken).ConfigureAwait(false);
+        });
+    }
+}
+
+static partial class MessageLogger
+{
+    /// <summary>
+    /// User not found: {<paramref name="userId"/>}
+    /// </summary>
+    [LoggerMessage(Level = LogLevel.Error, Message = "User not found: {userId}")]
+    public static partial void LogError_FailedToGetUser_UserNotFound(this ILogger logger, Guid userId);
+
+    /// <summary>
+    /// Failed to get a traQ user.
+    /// </summary>
+    [LoggerMessage(Level = LogLevel.Error, Message = "Failed to get a traQ user.")]
+    public static partial void LogError_FailedToGetUser(this ILogger logger, Exception exception);
+}
